requirements.txt
Django
sqlparse==0.5.0
psycopg2-binary>=2.9.6
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
djangorestframework-jwt==1.11.0
django-cors-headers==4.4.0
django-rest-authtoken==2.1.4
requests==2.32.3
tqdm==4.66.4
Werkzeug==3.1.3
django-extensions==3.2.3
gunicorn==23.0.0
boto3==1.35.81
django-storages==1.14.4
pandas
folium

